+++ packages/ocean-icons-react/src/Placeholder.tsx
@@ -18,7 +18,7 @@ const Placeholder = (
     {...props}
   >
     <path
-      stroke="#67697A"
+      stroke="currentColor"
       strokeLinecap="round"
       strokeLinejoin="round"
       strokeWidth={2}